Remove create_table_as_select helper from SqlClient interface
The underlying implementation of create_table_as_select is the same
across all engines, and we are moving MetricFlow away from being able
to issue write queries in general. This commit removes this particular
create table method in order to streamline the API and prevent new,
unintended uses of it from arising.

The remaining convenience wrappers will be removed from the SqlClient
interface when we move our test and tutorial data creation mechanisms
to rely on underlying dbt projects.