Move percentile rendering capability checks to rendering layer
Percentile aggregations are wildly inconsistent across SQL engines,
and so when they were first added we placed an engine capability
value in the SqlEngineAttributes struct. In hindsight, this was
inappropriate, since percentile support is really tied to the SQL
dialect as much as to the execution engine itself.

As we move to a model where we minimize engine-specific capability
constructs at the SqlClient level and instead focus on dialect
rendering differences, it makes sense to move these capability
markers to the expression renderer.