Remove SqlEngineAttributes struct from SqlClient protocol
The SqlClient protocol originally included a struct of engine properties which
was used to indicate certain engine-specific attributes which could be useful
to know about in various parts of the codebase where the SqlClient itself was
not required. The direct impetus of this construction was the need for a
backwards-compatible approach to an earlier implementation of the service
that would become MetricFlow.

As time has gone by, this property struct has become increasingly cumbersome,
serving as a living record of all of the engine-specific quirks we've ever
encountered, and conflating various dialect-specific rendering elements with
core engine capabilities and python API implementation quirks.

Now that we are narrowing MetricFlow's scope of responsibility for SqlClient
state and capability management, it has become clear that these engine-specific
property structs have outlived their usefulness. The majority of engine-specific
capability checks are either superfluous (e.g., FULL OUTER JOIN support is now
standard) or were rendered irrelevant (e.g., we no longer support isolation levels).
The remaining attributes are more appropriate as part of dialect rendering support,
leaving only the need for the SqlClient to hold pointers to the engine type and the
dialect renderer.