Update snapshots.
Note that derived metrics are handled with a different pattern than experssion
metrics. For cases where a expression metric is generated from 2 metrics from
a single data source, each one is computed separately and then joined. This
should be optimized via a case in the SourceScanOptimizer, but will be
handled as a follow up.