Clean up confusing test environment constructs
The dbt adapter-backed SqlClient needs to skip some tests and
depends on some custom package installations. This commit updates
the logic for the former to depend only on DuckDB, which is not
currently planned for migration to adapters. We also add a comment
explaining the installation configuration in the pyproject.toml.