Remove legacy MetricFlow config lookups from CLI tutorial
Once we build out the  dbt <-> MetricFLow integration, the MetricFlow
CLI will no longer need its own configuration file, and so we are
removing all references to it in favor of streamlining our CLI
connection to dbt project configs.

The tutorial command will eventually be updated to allow for
integration with dbt projects. Since we are working on this
in parallel this commit leaves the function largely intact,
but disables it in the CLI runtime and removes all problematic
references from the internal logic of the tutorial command itself.