Apply semantic manifest transformations to json parsed output
The dbt-generated semantic manifest is not transformed, and as such
it fails at query time with inscrutable errors about missing metrics
which clearly appear in all of the manifest artifacts and runtime
objects.

This change applies the semantic manifest transformations to all
dbt-sourced json-serialized semantic manifests, and embeds the call
inside of the dbt-specific function in order to prevent confusion.

Longer term, we should refine these APIs and unify them across
dbt-semantic-interfaces and MetricFlow so that we can always be
clear about whether or not transformations are needed for a given
serialized object. This change eschews the boolean parameter approach
because we only have one mode for json parsing at this time. If we
add a second approach, this would likely benefit from separate
methods rather than passing in an "apply_transformations" flag.