Integrate AdapterBackedSqlClient into CLI query command
The CLI was parsing semantic manifests from the dbt project root
but then falling back to MetricFlow's config-backed SqlClient
initialization, which meant users were effectively required to
match their legacy MetricFlow config with their dbt project profile
adapter configuration.

This change makes the dbt adapter available to the CLI, and also
sources the system schema element from the profile. This last is
likely temporary, as the only thing we seem to use it for is
testing and the time spine, the former of which can be moved to
test configuration constructs and the latter of which will be
moved into the semantic manifest itself.

As part of this change we consolidate the various dbt helpers into
a single location. Sorry about that, but it's annoying to revert.