Clean up semantic manifest builder
There were a couple of minor issues with the semantic manifest
builder function. It's also a bit odd to have it dangling
off of dbtArtifacts as a staticmethod - we might as well keep
everything built altogether in the dataclass. This commit
improves the integration of this class and fixes up the minor
issues.