Switch postgres test configuration to source from engine URL
Historically, MetricFlow has relied on SqlAlchemy engine URLs to
configure connections to its various engines, and currently all
testing is done via these URL connections.

In CI, in particular, we store secrets as these URL strings, and
essentially rely on them for configuring our engine-specific
test runners.

While we should move in the direction of consolidated shared
secret tokens with engine-specific extensions, rather than
jamming everything into a SqlAlchemy URL string and hoping
we never encounter problems with converting that to the format
dbt profiles expect, for the time being it's extremely
convenient to use the same environment configuration for
both engine types.

Therefore, we shift postgres setup to one where we parse the URL
and store the values in environment variables for use with the
dbt profile config.