Move Snowflake queries onto the dbt AdapterBackedSqlClient
It's time to enable Snowflake queries for our first class integration
with dbt projects.

This commit generalizes the initialization of the AdapterBackedSqlClient
to allow for easy expansion to other supported adapter types, and
demonstrates the capability by enabling support for Snowflake.

Due to Snowflake's insistence on storing database objects with fully
uppercase identifiers by default, and making queries against them
case sensitive..... sometimes? Maybe? .... we need to update a lot
of test assertion checks and a few of our accessors.

The only production impact is on the list_dimensions method, which
renders a column in a result dataframe that it selects by name. In
the longer term we likely need a better story around identifier
escaping for case sensitivity in Snowflake, but this appears to
work.

This commit was also tested via CLI queries against our Snowflake
instance, and everything works as expected.