Align postgres hatch environment config with Snowflake
Due to an incompatibility between dbt-snowflake 1.5 and 1.6, we
had to specify the version.

Given that versioning was necessary, the benefit of using pre-install
was limited. As such, we make the relevant adapters environment-specific
dependencies, and bring postgres in line. This will hopefully get
streamlined into the bundle package in the future, but for now
we can put up with manual version management on these adapter environments.