Fix tutorial loading with dbt project
During some live testing of the tutorial we discovered two seemingly
separate but closely connected issues. First, running the tutorial
from outside of a dbt project root filled the terminal with error
trace output, and second the dreaded ResourceWarning was appearing
after certain commands.

The base of the issue was due to us having to initialize the dbt project
at the very start of the CLI invocation, before any of our handlers
were active, and then accidentally initializing it a second time
as part of the tutorial operations.

This PR fixes both issues by first ensuring that we will simply not
run the CLI at all if we are not in a dbt project root, and along
the way restructuring our dbt artifact loading stack such that
we only invoke the dbtRunner one time per process.