Remove legacy tutorial helper functions
The original tutorial relied on dataframe-based table creation,
and cleaned up tables by calling drop_tables on the sql client.

We have switched to managing the tutorial datasets with dbt under
the hood, and we are pushing to deprecate the DDL operations within
the SqlClient interface, so we remove these legacy helpers in
order to cut down on callsites for the soon-to-be-removed SqlClient
methods.