Remove layer of indirection from CLI error messages
Currently, error messages in the CLI that result from failing
warehouse health checks or any problem with the model throw a
wrapped exception. In practice, this means a user with a misconfigured
model will see an error of the form:

`ERROR: An error occurred when attempting to create the MetricFlow engine`

along with a pointer to the log file, which will contain a much more
helpful error describing the root cause of the problem.

This change removes the wrapper so that the direct cause of the
exception will be surfaced in the CLI output itself, so users
will instead see exceptions that read something more like:

`ERROR: Metric `east_coast_order_amount` references measure `MeasureReference(element_name='orderss')` which has not been registered`