Move SqlRequestId and SqlTag classes to shared location
These classes were split between a module in the protocols package,
which is reserved for protocols, and a module in the sql_clients
package, which we are moving out of the production metricflow package
in order to eliminate concrete sql engine dependencies from the core
library.

This change combines these closely related modules into a new
package, named sql_request, and renames the old sql_request module
to the more descriptive sql_request_attributes.