Remove drop_table function from SqlClient protocol
This is only used in one place in production, and two places in tests,
one of which is a cleanup operation. More importantly, the only
implementations of the method were passing DROP TABLE IF EXISTS
statements into sql_client.execute(), so we simply inline those
calls in the handful of necessary locations.