Remove DDL methods from SqlClient protocol
MetricFlow is moving to a model where it serves as a library for
compiling and executing metric queries. This core functionality
does not require any DDL or other write operations, and so having
highly specific DDL and write query interfaces, such as the
create_table_from_dataframe method, is not warranted as part of the
core SqlClient protocol.

This PR removes those methods, and moves them to a test package
protocol in order to continue to support their only extant usage,
which is for test case setup, teardown, and verification around
certain operations.