Cut Redshift client over to dbt-redshift adapter
This removes the sqlalchemy-redshift/postgres implementation
of our redshift SqlClient class in favor of using the
AdapterBackedSqlClient with dbt-redshift as the underlying adapter.

Along with this change came a shift from psycopg2 to redshift-connector,
which is what the dbt adapter uses under the hood. This caused some updates
to types returned by redshift, necessitating a change to one of our
integration tests.

Finally, the dbt adpater appears to be materially slower than the
SqlAlchemy client, and on a smaller sized redshift cluster a random
subset of tests may fail with connection timeouts. The cause of this
is not clear, as configuration changes around timeouts, retries, and
auto-commit behavior made no discernable difference, and timings on
AdapterBackedSqlClient internal behaviors (e.g., converting agate to
pandas) suggest the issue is elsewhere.