Move Databricks query execution to AdapterBackedSqlClient
This effectively enables Databricks querying through the CLI for
the dbt-metricflow integation.

This includes a number of minor finicky databricks-specific adjustments
to the test classes and explain plan handling in the main class. These
are fairly hacky, but they are either specific to test classes and
therefore good enough until we can move to something more universally
robust like calling `dbt seed` or similar, or they are temporary until
some of the issues with the BaseAdapter.validate_sql_query method are
resolved.

As such, in order to allow for an all-green commit state, the legacy
SqlClient class will be removed in a follow-up.