Replace BigQuerySqlClient with AdapterBackedSqlClient instance
This enables BigQuery support for users of the dbt-metricflow
CLI integration.

The final cloud engine MetricFlow supports has now been moved to
the dbt adapter for CLI and test queries. BigQuery has a few
special features about it which added a bit of complexity to the
test configuration. BigQuery also requires us to use the new
validate_sql method on the adapter instance, as it does not
support the EXPLAIN keyword. The remaining adapters will follow
once we sort out handling for the macro substitution involved.

Due to a dependency version mismatch between dbt-bigquery and our
original SQLAlchemy-backed SqlClient implementation this commit
had to include the removal of the legacy BigQuery components.