Clean up dry run method in AdapterBackedSqlClient
The dry_run method in AdapterBackedSqlClient is a bit messy due to
some snags with using the SQLAdapter implementation of dbt's
BaseAdapter.validate_sql method.

Once those issues are resolved we can consolidate to a single call
type, which will require a connection with a standardized identifier.

In the interim, we can consolidate the SQLAdapter subclass behavior
onto a single named connection invocation and simply handle the results
in a special way if that adapter happens to be Databricks.