Improve messaging for populating the persistent source schema
When the test runner encounters an error with the persistent source
schema it prompts the user with an example command. That command
no longer works, and it was very difficult to find in the sea
of log output for a failed test run even with the promotion to a
warning.

This change makes the command work in an engine-specific way, which
is the current approach to our test environment management, and
improves the formatting a bit to make it more copy/paste friendly.