Always render granularity in time dimension names
The original implementation of MetricFlow treated DAY granularity
as being particularly special, in part because it's the most common
query granularity and in part because it is the minimum granularity
we currently support.

The special treatment essentially made any time dimension queried
with DAY granularity render as the bare name of the dimension,
regardless of that dimension's base granularity. This is a bit odd,
because "billing_month" would only render as "billing_month" if the
granularity was DAY, but in the normal usage of it it would render
instead as "billing_month__month"

Since we plan to support broader granularities, and the inconsistent
output is confusing, this change makes all time dimensions render
with the granularity suffix across our query rendering and output
data column naming. We do this by removing the special casing around
dimension naming for DAY granularity.