Update dependencies and attribution listing
In preparation for launching 0.200.0, we update the attribution.md
to be more up to date with our current dependency suite. This
initial run exposed some packages which we are not currently using,
and a handful of other dependency versions which we should update
in our dev builds, so we make the relevant changes in order to
do a final attribution and license analysis prior to launch.