Make duckdb the standard in the dev-env environment
MetricFlow's testing configuration currently tags sql engine types
to hatch execution environment. The standard one, which we use for
`make test`, is meant to be duckdb, but a stale environment variable
can cause inscrutable failures on any test requiring an engine
connection.

This change updates the test environment configuration to make duckdb
the standard for any runs in the hatch dev-env, including any
invocation such as `make test`. It also removes the default fallback
to duckdb, which was causing inconsistent behaviors in certain
environment configurations, and replaces it with a consistent error.

Developers with heavily customized environment configurations might
have a bad time with this, but they were already having a bad time
with other things, and at least this gives them an actionable
error message.