Use year instead of isoyear for BigQuery date_trunc operations
BigQuery's time granularity truncation implementation relied,
correctly, on ISOWEEK for weekly granularities (to have the week
start on Monday) and, incorrectly, on ISOYEAR for yearly granularities
(which made the year start on whatever Monday happened to be for the
week containing the first Thursday of the year).

Every other engine we support renders date_trunc to truncate to
the first of January for yearly granularities.

Before this change, BigQuery would adjust the granularity for
the date 2015-06-04 to 2014-12-29

After this change, BigQuery will adjust the granularity for
the date 2015-06-04 to 2015-01-01

This is consistent with the behavior of all other supported engines.