Allow tabulate versions >= 0.8.9
Tabulate is an optional dependency of Pandas, which we use for printing
our query results. This means we must depend on tabulate directly
otherwise users see unhelpful errors about installing the tabulate
dependency.

Our version was pinned to ~0.8.9, but this comes with unexpected
limitations for users interacting with other packages in the broader
dbt ecosystem which require 0.9 or later, so here we are.