Remove generic source dataset constructs
The original implementation of MetricFlow used a generic type
annotation for the "Source DataSet" of the dataflow plan. This was
in place to support some specific implementations of Transform's
proprietary product logic, and as an interface boundary it proved
ineffective. Consequently, attempts to revisit that functionality
will rely on other approaches.

Within MetricFlow, this generic type resolves, to one of two concrete types:
the SqlDataSet, or the SemanticModelDataSet, which inherits from
SqlDataSet. Since it's far simpler to just use SqlDataSet as the
required type than to litter the codebase with generic typehints and
concrete type specs in some places but not others, we make that change here.

Sadly, this must all be done in one massive commit, because there's no way
to get the typechecker to pass by removing this one stage at a time without
doing considerably more restructuring around the data set handoffs.

This commit was produced via a mixture of mechanical find/replace
approaches within VSCode, along with file-by-file manual fixups for
cases where leftovers were causing trouble.