Remove unused grain_to_date from SqlTimeDeltaExpression
The original implementation of SqlTimeDeltaExpression had a custom
override for grain_to_date intended to support cumulative metrics.
This parameter would change the expression from a time delta to a
date_trunc. At some point we cleaned up the callsites to invoke
date_trunc directly instead of offloading this work to an overloaded
class.

This commit simply removes the confusing parameter in order to streamline
the expression rendering for time delta operations.