Rename SqlTimeDeltaExpression to SqlSubtractTimeIntervalExpression
The original name for this class was misleading, as a time delta is
effectively an interval computed as a difference. This suggested either
that the rendering should be a date_diff to produce an interval between
two timestamps, or that it should be an interval expression for use
in some other operation.

In reality, this class provides the functionality of a date_subtract,
where we subtract an interval value from a given input timestamp.