Add Trino adapter support for MetricFlow
This commit, authored by @sarbmeetka, includes all of the material
changes relevant for adding support for Trino to MetricFlow.

Specifically, this includes the SqlQueryPlanRenderer for Trino and
integration in the CLI and unit test suite with the dbt-trino
adapter. All other changes - particularly those to the contents
of test queries - were made in order to bring our internal test
SQL in line with Trino's ANSI-compliance requirements.

The original PR (https://github.com/dbt-labs/metricflow/pull/810)
commit stack has been consolidated into this one, and a separate
commit with the auto-generated snapshot files for the Trino test
cases. This has been done to make isolating fixes a bit simpler
for CI testing prior to merge.