Add tests to ensure consistent results for SqlExtractExpression
Our initial date_part implementation essentially assumed all engines
had the same behavior for various date part extraction calls, but this
is not the case, as things like day of week and year can vary from
engine to engine depending on whether or not they hew to the ISO
standards for such things.

This commit adds some engine tests for SqlExtractExpression so we
can track and enforce consistency across engine types.