Remove support for date_part `week` from extract expression
Support for date_part `week` is fraught with peril, as there is
no good standard way of extracting a week number from a date.

Because we hew to the calendar standard for date_part `year`
(as opposed to using ISO year, which defines the year start as
the Monday of the week containing the first Thursday in January)
it would be odd to use ISO week of year, which is based on the
ISO year. However, no other standard exists, and so we simply
remove the property from the enum and thereby drop support for
it. We can add it later and apply a well-defined behavior to it,
if necessary.

For more details on the mayhem please refer to the docblock
changes for the DatePart enum that accompany this commit.