Consolidate dataflow_to_sql tests within their module
The test_dataflow_to_sql_plan module has become a dumping
ground for a mixture of direct tests of DataflowPlan to
SqlQueryPlan object conversion and tests which are more about
query rendering for a variety of metric queries.

In preparation for splitting the latter class of tests into
their own location, we consolidate the former class into
a single block at the top of the module for ease of review.