Remove capsys.disabled magic from CLI tests
The addition of capsys.disabled calls to the CLI tests was done
to work around an issue with files being opened and closed in the
CLI setup fixtures. When running under high parallelism the CLI
tests would occasionally be split up across worker instances, and
so the file system operations - which were effectively creating
and removing the same file - would cause failures due to race
conditions between the different test runner fixture executions.

Adding capsys.disabled didn't solve this problem, it merely made
it somewhat less likely to occur, and ultimately replaced the error
with a different one (file not found).

This commit removes the capsys.disabled calls in favor of a
follow-up to stop doing all of this filesystem manipulation.