Expand cumulative metric test coverage for non-adjustable time filters
Cumulative metric query rendering currently follows the same
pattern as all other metrics, namely, we join -> filter -> aggregate.

The problem is, if the filter specifies a restrictive time span, we
may apply the filter on the inputs to the aggregations, which will
produce incorrect cumulative metric outputs.

This will happen in any case where the dimension `metric_time` is
not incorporated into the query, as illustrated by the changes to the
plan files for test_cumulative_metric_with_time_constraint.

For now, we simply illustrate current behavior with the test cases
in preparation for adding a validation enhancement to prevent bad
queries from being processed. We will expand support for cumulative
metric queries beyond metric_time in a follow-up.