Update mypy to ~=1.7.0
New errors with type: ignore specifications have cropped up
in a transitive dependency. It turns out the specific type: ignore
spec used in that library is supported in mypy 1.7, but not 1.3.

Note mypy does not use SemVer, as minor releases may add functionality
that will detect latent issues in code, so we will continue with
periodic manual updates.