Add query issue classes.
Previously, the query parser would raise an exception for a single issue
discovered during query parsing. To aid debugging, we're changing this behavior
to return all errors together so that the user can fix all issues in a single
pass. This commit adds classes that can represent those issues and also adds a
way of encapsulating issues into groups.