Add GroupByItemResolver.
The GroupByItemResolver is a class that helps to resolve patterns into concrete
specs based on the structure of the query. Different use cases affect how
patterns describing ambiguous inputs are resolved - e.g. as a query input, a
time dimension without a grain is resolved to the finest time grain available.
In where filters, a time dimension without a grain can only be resolved if all
parents have the same defined grain.