Add query rendering test for filters on nested derived metrics
We received a report of filters on nested derived metrics failing
to render. This issue was resolved as part of the updates we made
to our group by item and filter resolution components, and should
no longer be a problem with the new query parsing and dataflow plan
layout components. However, those components did not include a query
rendering test case for nested derived filters.

This adds the relevant test case from query specification through
the query parser and final SQL rendering.