Remove default time constraint for queries with cumulative metrics.
A previous change (https://github.com/dbt-labs/metricflow/pull/918) updated
the logic to apply a default time constraint for queries with cumulative metrics
if one was not supplied. However, the fix revealed that the feature was not
working for some time, possibly since the public release. I also realized that
the original logic (and also the fix) is not correct with derived metrics that
use cumulative metrics. There is future work to better handle time constraints /
predicate push down, so removing this for now.