Update local postgres setup instructions
The original setup instructions for local configuration of a
postgres instance for testing pre-dated our cutover to hatch for
environment management. Consequently, there were a number of
inaccuracies in how to use the postgres container once it was up
and running. This commit brings those command examples in line with
our current configuration.